+++ libnavigation-core/src/test/java/com/mapbox/navigation/core/MapboxNavigationTest.kt
@@ -1,6 +1,7 @@
 package com.mapbox.navigation.core
 
 import com.mapbox.android.telemetry.TelemetryEnabler
+import com.mapbox.api.directions.v5.models.DirectionsRoute
 import com.mapbox.api.directions.v5.models.RouteOptions
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.base.options.IncidentsOptions
@@ -23,7 +24,8 @@ import com.mapbox.navigation.core.reroute.RerouteController
 import com.mapbox.navigation.core.reroute.RerouteState
 import com.mapbox.navigation.core.telemetry.MapboxNavigationTelemetry
 import com.mapbox.navigation.core.trip.session.LocationObserver
-import com.mapbox.navigation.core.trip.session.NativeSetRouteResult
+import com.mapbox.navigation.core.trip.session.NativeSetRouteError
+import com.mapbox.navigation.core.trip.session.NativeSetRouteValue
 import com.mapbox.navigation.core.trip.session.NavigationSession
 import com.mapbox.navigation.core.trip.session.OffRouteObserver
 import com.mapbox.navigation.core.trip.session.RoadObjectsOnRouteObserver
@@ -31,6 +33,7 @@ import com.mapbox.navigation.core.trip.session.TripSessionState
 import com.mapbox.navigation.core.trip.session.createSetRouteResult
 import com.mapbox.navigation.testing.factories.createDirectionsRoute
 import com.mapbox.navigation.testing.factories.createNavigationRoute
+import com.mapbox.navigation.testing.factories.createRouteOptions
 import com.mapbox.navigator.FallbackVersionsObserver
 import com.mapbox.navigator.NavigatorConfig
 import com.mapbox.navigator.RouteAlternative
@@ -256,6 +259,19 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
     }
 
     @Test
+    fun onDestroyDoesNotSetRoutesToEmptyIfEmptyIsInvalid() = coroutineRule.runBlockingTest {
+        createMapboxNavigation()
+        coEvery {
+            tripSession.setRoutes(any(), any(), any())
+        } returns NativeSetRouteError("some error")
+        mapboxNavigation.onDestroy()
+
+        verify(exactly = 0) {
+            directionsSession.setRoutes(any(), any(), any())
+        }
+    }
+
+    @Test
     fun onDestroyCallsTripSessionStop() {
         createMapboxNavigation()
         mapboxNavigation.onDestroy()
@@ -395,6 +411,152 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
         verify(ordering = Ordering.ORDERED) {
             tripSession.registerOffRouteObserver(any())
             rerouteController.reroute(any())
+        }
+    }
+
+    @Test
+    fun `new routes are set after reroute`() {
+        val newRoutes = listOf(mockk<NavigationRoute>(), mockk())
+        val navigationRerouteController: NavigationRerouteController? = mockk(relaxed = true) {
+            every { reroute(any<NavigationRerouteController.RoutesCallback>()) } answers {
+                (firstArg() as NavigationRerouteController.RoutesCallback)
+                    .onNewRoutes(newRoutes, mockk())
+            }
+        }
+        coEvery {
+            tripSession.setRoutes(any(), any(), any())
+        } returns NativeSetRouteValue(emptyList())
+
+        createMapboxNavigation()
+        mapboxNavigation.setRerouteController(navigationRerouteController)
+        val observers = mutableListOf<OffRouteObserver>()
+        verify { tripSession.registerOffRouteObserver(capture(observers)) }
+
+        observers.forEach {
+            it.onOffRouteStateChanged(true)
+        }
+        coVerify(exactly = 1) {
+            directionsSession.setRoutes(newRoutes, 0, RoutesExtra.ROUTES_UPDATE_REASON_REROUTE)
+        }
+    }
+
+    @Test
+    fun `new routes are not set after reroute if they are invalid`() {
+        val newRoutes = listOf(mockk<NavigationRoute>(), mockk())
+        val navigationRerouteController: NavigationRerouteController? = mockk(relaxed = true) {
+            every { reroute(any<NavigationRerouteController.RoutesCallback>()) } answers {
+                (firstArg() as NavigationRerouteController.RoutesCallback)
+                    .onNewRoutes(newRoutes, mockk())
+            }
+        }
+        coEvery {
+            tripSession.setRoutes(any(), any(), any())
+        } returns NativeSetRouteError("some error")
+
+        createMapboxNavigation()
+        mapboxNavigation.setRerouteController(navigationRerouteController)
+        val observers = mutableListOf<OffRouteObserver>()
+        verify { tripSession.registerOffRouteObserver(capture(observers)) }
+
+        observers.forEach {
+            it.onOffRouteStateChanged(true)
+        }
+        coVerify(exactly = 0) {
+            directionsSession.setRoutes(any(), any(), any())
+        }
+    }
+
+    @Test
+    fun `set reroute controller in fetching state sets routes to session`() {
+        val newRoutes = listOf(mockk<NavigationRoute>(), mockk())
+        val oldController = mockk<RerouteController>(relaxed = true) {
+            every { state } returns RerouteState.FetchingRoute
+        }
+        val navigationRerouteController: NavigationRerouteController? = mockk(relaxed = true) {
+            every { reroute(any<NavigationRerouteController.RoutesCallback>()) } answers {
+                (firstArg() as NavigationRerouteController.RoutesCallback)
+                    .onNewRoutes(newRoutes, mockk())
+            }
+        }
+        coEvery {
+            tripSession.setRoutes(any(), any(), any())
+        } returns NativeSetRouteValue(emptyList())
+
+        createMapboxNavigation()
+        mapboxNavigation.setRerouteController(oldController)
+        mapboxNavigation.setRerouteController(navigationRerouteController)
+        coVerify(exactly = 1) {
+            directionsSession.setRoutes(newRoutes, 0, RoutesExtra.ROUTES_UPDATE_REASON_REROUTE)
+        }
+    }
+
+    @Test
+    fun `set reroute controller in fetching state does not set invalid routes to session`() {
+        val newRoutes = listOf(mockk<NavigationRoute>(), mockk())
+        val oldController = mockk<RerouteController>(relaxed = true) {
+            every { state } returns RerouteState.FetchingRoute
+        }
+        val navigationRerouteController: NavigationRerouteController? = mockk(relaxed = true) {
+            every { reroute(any<NavigationRerouteController.RoutesCallback>()) } answers {
+                (firstArg() as NavigationRerouteController.RoutesCallback)
+                    .onNewRoutes(newRoutes, mockk())
+            }
+        }
+        coEvery {
+            tripSession.setRoutes(any(), any(), any())
+        } returns NativeSetRouteError("some error")
+
+        createMapboxNavigation()
+        mapboxNavigation.setRerouteController(oldController)
+        mapboxNavigation.setRerouteController(navigationRerouteController)
+        coVerify(exactly = 0) {
+            directionsSession.setRoutes(any(), any(), any())
+        }
+    }
+
+    @Test
+    fun `set legacy reroute controller in fetching state sets routes to session`() {
+        val newRoutes = emptyList<DirectionsRoute>()
+        val oldController = mockk<RerouteController>(relaxed = true) {
+            every { state } returns RerouteState.FetchingRoute
+        }
+        val rerouteController: RerouteController = mockk(relaxed = true) {
+            every { reroute(any()) } answers {
+                (firstArg() as RerouteController.RoutesCallback).onNewRoutes(newRoutes)
+            }
+        }
+        coEvery {
+            tripSession.setRoutes(any(), any(), any())
+        } returns NativeSetRouteValue(emptyList())
+
+        createMapboxNavigation()
+        mapboxNavigation.setRerouteController(oldController)
+        mapboxNavigation.setRerouteController(rerouteController)
+        coVerify(exactly = 1) {
+            directionsSession.setRoutes(any(), 0, RoutesExtra.ROUTES_UPDATE_REASON_REROUTE)
+        }
+    }
+
+    @Test
+    fun `set legacy reroute controller in fetching state does not set invalid routes to session`() {
+        val newRoutes = emptyList<DirectionsRoute>()
+        val oldController = mockk<RerouteController>(relaxed = true) {
+            every { state } returns RerouteState.FetchingRoute
+        }
+        val rerouteController: RerouteController = mockk(relaxed = true) {
+            every { reroute(any()) } answers {
+                (firstArg() as RerouteController.RoutesCallback).onNewRoutes(newRoutes)
+            }
+        }
+        coEvery {
+            tripSession.setRoutes(any(), any(), any())
+        } returns NativeSetRouteError("some error")
+
+        createMapboxNavigation()
+        mapboxNavigation.setRerouteController(oldController)
+        mapboxNavigation.setRerouteController(rerouteController)
+        coVerify(exactly = 0) {
+            directionsSession.setRoutes(any(), any(), any())
         }
     }
 
@@ -620,16 +782,17 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
     fun `setRoute pushes the route to the directions session`() = coroutineRule.runBlockingTest {
         createMapboxNavigation()
         val route: NavigationRoute = mockk()
-        val routeOptions: RouteOptions = mockk()
+        val routeOptions = createRouteOptions()
         every { route.routeOptions } returns routeOptions
         every { route.directionsRoute.geometry() } returns "geometry"
         every { route.directionsRoute.legs() } returns emptyList()
-        every { routeOptions.overview() } returns "full"
-        every { routeOptions.annotationsList() } returns emptyList()
 
         val routes = listOf(route)
         val initialLegIndex = 2
 
+        coEvery {
+            tripSession.setRoutes(routes, initialLegIndex, RoutesExtra.ROUTES_UPDATE_REASON_NEW)
+        } returns NativeSetRouteValue(emptyList())
         mapboxNavigation.setNavigationRoutes(routes, initialLegIndex)
 
         verify(exactly = 1) {
@@ -638,6 +801,111 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
             )
         }
     }
+
+    @Test
+    fun `setRoute does not push the invalid route to the directions session`() =
+        coroutineRule.runBlockingTest {
+            createMapboxNavigation()
+            val route: NavigationRoute = mockk()
+            val routeOptions = createRouteOptions()
+            every { route.routeOptions } returns routeOptions
+            every { route.directionsRoute.geometry() } returns "geometry"
+            every { route.directionsRoute.legs() } returns emptyList()
+
+            val routes = listOf(route)
+            val initialLegIndex = 2
+
+            coEvery {
+                tripSession.setRoutes(routes, initialLegIndex, RoutesExtra.ROUTES_UPDATE_REASON_NEW)
+            } returns NativeSetRouteError("some error")
+            mapboxNavigation.setNavigationRoutes(routes, initialLegIndex)
+
+            verify(exactly = 0) {
+                directionsSession.setRoutes(any(), any(), any())
+            }
+        }
+
+    @Test
+    fun `setNavigationRoutes calls callback with successfully set routes`() =
+        coroutineRule.runBlockingTest {
+            createMapboxNavigation()
+            val routes = listOf(mockk<NavigationRoute>(relaxed = true))
+            val callback = mockk<RoutesSetCallback>(relaxed = true)
+            val initialLegIndex = 2
+
+            coEvery {
+                tripSession.setRoutes(routes, initialLegIndex, RoutesExtra.ROUTES_UPDATE_REASON_NEW)
+            } returns NativeSetRouteValue(emptyList())
+            mapboxNavigation.setNavigationRoutes(routes, initialLegIndex, callback)
+
+            val successes = mutableListOf<RoutesSetCallbackSuccess>()
+            verify(exactly = 1) { callback.onRoutesSetResult(capture(successes)) }
+            verify(exactly = 0) { callback.onRoutesSetError(any()) }
+
+            successes[0].run {
+                assertEquals(routes, this.routes)
+            }
+        }
+
+    @Test
+    fun `setNavigationRoutes calls callback with error for invalid routes`() =
+        coroutineRule.runBlockingTest {
+            createMapboxNavigation()
+            val routes = listOf(mockk<NavigationRoute>(relaxed = true))
+            val callback = mockk<RoutesSetCallback>(relaxed = true)
+            val initialLegIndex = 2
+            val error = "some error"
+
+            coEvery {
+                tripSession.setRoutes(routes, initialLegIndex, RoutesExtra.ROUTES_UPDATE_REASON_NEW)
+            } returns NativeSetRouteError(error)
+            mapboxNavigation.setNavigationRoutes(routes, initialLegIndex, callback)
+
+            val errors = mutableListOf<RoutesSetCallbackError>()
+            verify(exactly = 1) { callback.onRoutesSetError(capture(errors)) }
+            verify(exactly = 0) { callback.onRoutesSetResult(any()) }
+
+            errors[0].run {
+                assertEquals(routes, this.routes)
+                assertEquals(error, this.error)
+            }
+        }
+
+    @Test
+    fun `deprecated setRoutes pushes the route to the directions session`() =
+        coroutineRule.runBlockingTest {
+            createMapboxNavigation()
+
+            val routes = emptyList<DirectionsRoute>()
+            val initialLegIndex = 2
+
+            coEvery {
+                tripSession.setRoutes(any(), any(), any())
+            } returns NativeSetRouteValue(emptyList())
+            mapboxNavigation.setRoutes(routes, initialLegIndex)
+
+            verify(exactly = 1) {
+                directionsSession.setRoutes(any(), any(), any())
+            }
+        }
+
+    @Test
+    fun `deprecated setRoutes does not push the invalid route to the directions session`() =
+        coroutineRule.runBlockingTest {
+            createMapboxNavigation()
+
+            val routes = emptyList<DirectionsRoute>()
+            val initialLegIndex = 2
+
+            coEvery {
+                tripSession.setRoutes(any(), any(), any())
+            } returns NativeSetRouteError("some error")
+            mapboxNavigation.setRoutes(routes, initialLegIndex)
+
+            verify(exactly = 0) {
+                directionsSession.setRoutes(any(), any(), any())
+            }
+        }
 
     @Test
     fun `requestRoutes pushes the request to the directions session`() {
@@ -965,11 +1233,11 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
         val shortRoutes = listOf<NavigationRoute>(mockk())
         coEvery { tripSession.setRoutes(longRoutes, any(), any()) } coAnswers {
             delay(100L)
-            mockk(relaxed = true)
+            NativeSetRouteValue(emptyList())
         }
         coEvery { tripSession.setRoutes(shortRoutes, any(), any()) } coAnswers {
             delay(50L)
-            mockk(relaxed = true)
+            NativeSetRouteValue(emptyList())
         }
 
         pauseDispatcher {
@@ -984,43 +1252,44 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
     }
 
     @Test
-    fun `route refresh of previous route completes after new route is set`() = coroutineRule.runBlockingTest {
-        every { NavigationComponentProvider.createDirectionsSession(any()) } answers {
-            MapboxDirectionsSession(mockk(relaxed = true))
-        }
-        createMapboxNavigation()
-        val first = listOf(createNavigationRoute(createDirectionsRoute(requestUuid = "test1")))
-        val second = listOf(createNavigationRoute(createDirectionsRoute(requestUuid = "test2")))
-        val refreshOrFirstRoute = CompletableDeferred<Unit>()
-        coEvery { routeRefreshController.refresh(any()) } coAnswers {
-            CompletableDeferred<Unit>().await() // never completes
-            firstArg()
-        }
-        coEvery { routeRefreshController.refresh(first) } coAnswers {
-            refreshOrFirstRoute.await()
-            listOf(createNavigationRoute(createDirectionsRoute(requestUuid = "test1.1")))
-        }
-        coEvery { tripSession.setRoutes(second, any(), any()) } coAnswers {
-            mockk(relaxed = true)
-        }
+    fun `route refresh of previous route completes after new route is set`() =
+        coroutineRule.runBlockingTest {
+            every { NavigationComponentProvider.createDirectionsSession(any()) } answers {
+                MapboxDirectionsSession(mockk(relaxed = true))
+            }
+            createMapboxNavigation()
+            val first = listOf(createNavigationRoute(createDirectionsRoute(requestUuid = "test1")))
+            val second = listOf(createNavigationRoute(createDirectionsRoute(requestUuid = "test2")))
+            val refreshOrFirstRoute = CompletableDeferred<Unit>()
+            coEvery { routeRefreshController.refresh(any()) } coAnswers {
+                CompletableDeferred<Unit>().await() // never completes
+                firstArg()
+            }
+            coEvery { routeRefreshController.refresh(first) } coAnswers {
+                refreshOrFirstRoute.await()
+                listOf(createNavigationRoute(createDirectionsRoute(requestUuid = "test1.1")))
+            }
+            coEvery { tripSession.setRoutes(second, any(), any()) } coAnswers {
+                NativeSetRouteValue(emptyList())
+            }
 
-        val routesUpdates = mutableListOf<RoutesUpdatedResult>()
-        mapboxNavigation.registerRoutesObserver {
-            routesUpdates.add(it)
-        }
-        mapboxNavigation.setNavigationRoutes(first)
-        mapboxNavigation.setNavigationRoutes(second)
-        refreshOrFirstRoute.complete(Unit)
+            val routesUpdates = mutableListOf<RoutesUpdatedResult>()
+            mapboxNavigation.registerRoutesObserver {
+                routesUpdates.add(it)
+            }
+            mapboxNavigation.setNavigationRoutes(first)
+            mapboxNavigation.setNavigationRoutes(second)
+            refreshOrFirstRoute.complete(Unit)
 
-        assertEquals(
-            listOf(first, second),
-            routesUpdates.map { it.navigationRoutes }
-        )
-        assertEquals(
-            listOf(RoutesExtra.ROUTES_UPDATE_REASON_NEW, RoutesExtra.ROUTES_UPDATE_REASON_NEW),
-            routesUpdates.map { it.reason }
-        )
-    }
+            assertEquals(
+                listOf(first, second),
+                routesUpdates.map { it.navigationRoutes }
+            )
+            assertEquals(
+                listOf(RoutesExtra.ROUTES_UPDATE_REASON_NEW, RoutesExtra.ROUTES_UPDATE_REASON_NEW),
+                routesUpdates.map { it.reason }
+            )
+        }
 
     @Test
     fun `set route - immediately stops the reroute controller`() = coroutineRule.runBlockingTest {
@@ -1030,7 +1299,7 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
         val shortRoutes = listOf<NavigationRoute>(mockk())
         coEvery { tripSession.setRoutes(shortRoutes, any(), any()) } coAnswers {
             delay(50L)
-            mockk(relaxed = true)
+            NativeSetRouteValue(emptyList())
         }
 
         pauseDispatcher {
@@ -1047,7 +1316,7 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
             val nativeAlternatives = listOf<RouteAlternative>(mockk())
             coEvery {
                 tripSession.setRoutes(routes, 0, RoutesExtra.ROUTES_UPDATE_REASON_NEW)
-            } returns NativeSetRouteResult(nativeAlternatives)
+            } returns NativeSetRouteValue(nativeAlternatives)
 
             mapboxNavigation.setNavigationRoutes(routes)
 
@@ -1081,7 +1350,7 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
                     initialLegIndex,
                     RoutesExtra.ROUTES_UPDATE_REASON_REFRESH
                 )
-            } returns NativeSetRouteResult()
+            } returns NativeSetRouteError("some error")
             coEvery { routeRefreshController.refresh(routes) } returns refreshedRoutes
 
             verify { directionsSession.registerRoutesObserver(capture(routeObserversSlot)) }
@@ -1107,7 +1376,7 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
                 tripSession.setRoutes(routes, 0, RoutesExtra.ROUTES_UPDATE_REASON_NEW)
             } coAnswers {
                 delay(100)
-                mockk(relaxed = true)
+                NativeSetRouteValue(emptyList())
             }
             every { directionsSession.setRoutes(any(), any(), any()) } answers {
                 every { directionsSession.routes } returns firstArg()
@@ -1187,6 +1456,42 @@ internal class MapboxNavigationTest : MapboxNavigationBaseTest() {
                     0,
                     RoutesExtra.ROUTES_UPDATE_REASON_REFRESH
                 )
+            }
+        }
+
+    @Test
+    fun `refreshed route is not set to directions session if it is invalid`() =
+        coroutineRule.runBlockingTest {
+            createMapboxNavigation()
+            val primary: NavigationRoute = mockk {
+                every { directionsRoute } returns mockk()
+            }
+            val routes = listOf(primary)
+            val reason = RoutesExtra.ROUTES_UPDATE_REASON_NEW
+            val initialLegIndex = 0
+            val routeObserversSlot = mutableListOf<RoutesObserver>()
+            every { tripSession.getState() } returns TripSessionState.STARTED
+            every { directionsSession.initialLegIndex } returns initialLegIndex
+            verify { directionsSession.registerRoutesObserver(capture(routeObserversSlot)) }
+
+            val refreshedRoutes = listOf(mockk<NavigationRoute>())
+            coEvery { routeRefreshController.refresh(routes) } returns refreshedRoutes
+            coEvery {
+                tripSession.setRoutes(any(), any(), any())
+            } returns NativeSetRouteError("some error")
+            routeObserversSlot.forEach {
+                it.onRoutesChanged(RoutesUpdatedResult(routes, reason))
+            }
+
+            coVerify(exactly = 1) {
+                tripSession.setRoutes(
+                    refreshedRoutes,
+                    0,
+                    RoutesExtra.ROUTES_UPDATE_REASON_REFRESH
+                )
+            }
+            verify(exactly = 0) {
+                directionsSession.setRoutes(any(), any(), any())
             }
         }